pyproject.toml
[project]
name = "tgmusicbot"
version = "1.2.4"
description = "Telegram Music Bot with full playback support using PyTdBot and PyTgCalls"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Ashok Shau", email = "ashokshau@t.me" },
]
keywords = [
    "telegram",
    "music",
    "bot",
    "async",
    "tdlib",
    "pytgcalls",
    "voicechat",
    "spotify",
    "ytmusic",
    "opensource",
    "pytdbot",
    "py-tgcalls",
    "yt-dlp",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat :: Telegram",
    "Topic :: Multimedia :: Sound/Audio :: Players",
    "Environment :: Console",
    "Framework :: AsyncIO",
]
dependencies = [
    "aiofiles>=24.1.0",
    "cachetools>=6.2.0",
    "kurigram>=2.2.12",
    "ntgcalls>=2.0.6",
    "pillow>=11.3.0",
    "psutil>=7.1.0",
    "py-tgcalls>=2.2.8",
    "py-yt-search>=0.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.9",
    "pymongo>=4.15.1",
    "pytdbot[tdjson]>=0.9.6",
    "pytgcrypto>=1.2.11",
    "ujson>=5.11.0",
    "yt-dlp>=2025.9.5",
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
dev = [
    "black",
]

[project.scripts]
start = "TgMusic.__main__:main"

[project.urls]
Homepage = "https://github.com/AshokShau/tgmusicbot"
Source = "https://github.com/AshokShau/tgmusicbot"
"Issue Tracker" = "https://github.com/AshokShau/tgmusicbot/issues"

[tool.uv]
package = true

[tool.uv.sources]
pytdbot = { git = "https://github.com/pytdbot/client.git" }

[tool.setuptools]
packages = [
    "TgMusic",
]
